backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
sentence-transformers==2.3.1
faiss-cpu==1.8.0
numpy==1.26.3
PyPDF2==3.0.1
aiofiles==23.2.1
